requirements.txt
# Core LLM Providers
openai
groq
google-generativeai
mistralai

# LangChain Ecosystem
langchain
langchain-openai
langchain-community
langchain-core
langchain-google-genai
langchain_huggingface

# Vector Store
faiss-cpu
chromadb

# Tokenizer
tiktoken

# API Framework (if building backend)
fastapi
uvicorn

# Environment Variables
python-dotenv

# Data Handling
pydantic
requests

# Optional (for RAG / document loaders)
pypdf
beautifulsoup4
lxml

# Optional (for Jupyter development)
notebook
ipykernel